gallery/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-environ==0.4.5
Pillow==9.3.0
pytz==2022.6
sqlparse==0.4.3
